feat: split component to BackupTables, SnapshotTables, and ConfirmActionForm components for enhanced backup management UI
