+++ share/dashboard_react/src/Pages/Maintenance/components/BackupTables.jsx
@@ -1,0 +1,147 @@
+import React, { useMemo } from 'react'
+import { createColumnHelper } from '@tanstack/react-table'
+import { Box, VStack } from '@chakra-ui/react'
+import { DataTable } from '../../../components/DataTable'
+import TableType3 from '../../../components/TableType3'
+import { formatBytes, formatDate, getBackupMethod, getBackupStrategy } from '../../../utility/common'
+import styles from '../styles.module.scss'
+
+const columnHelper = createColumnHelper()
+
+function BackupTables({ data, backupStats }) {
+  const columns = useMemo(
+    () => [
+      columnHelper.accessor((row) => row.id, {
+        cell: (info) => info.getValue(),
+        header: 'ID',
+        id: 'id'
+      }),
+      columnHelper.accessor(
+        (row) => (
+          <>
+            {formatDate(row.startTime)} <br />
+            {formatDate(row.endTime)}
+          </>
+        ),
+        {
+          cell: (info) => info.getValue(),
+          header: 'Start - End Time',
+          id: 'startendTime',
+          minWidth: 160
+        }
+      ),
+      columnHelper.accessor(
+        (row) => (
+          <VStack className={styles.cellStack}>
+            <Box className={styles.cellValue}>{getBackupMethod(row.backupMethod)}</Box>
+            <Box className={styles.cellValue}>{row.backupTool}</Box>
+          </VStack>
+        ),
+        {
+          cell: (info) => info.getValue(),
+          header: 'Backup Method / Tool',
+          id: 'backupMethod'
+        }
+      ),
+      columnHelper.accessor((row) => getBackupStrategy(row.backupStrategy), {
+        cell: (info) => info.getValue(),
+        header: 'Strategy',
+        id: 'strategy'
+      }),
+      columnHelper.accessor(
+        (row) => (
+          <VStack className={styles.cellStack}>
+            <Box className={styles.cellValue}>{row.source}</Box>
+            <Box className={styles.cellValue}>{row.dest}</Box>
+          </VStack>
+        ),
+        {
+          cell: (info) => info.getValue(),
+          header: 'Source - Dest',
+          id: 'srcDest'
+        }
+      ),
+      columnHelper.accessor((row) => formatBytes(row.size), {
+        cell: (info) => info.getValue(),
+        header: 'Backup Size',
+        id: 'backupSize',
+        minWidth: 100
+      }),
+      columnHelper.accessor((row) => (row.compressed ? 'Yes' : 'No'), {
+        cell: (info) => info.getValue(),
+        header: 'Compressed',
+        id: 'compression'
+      }),
+      columnHelper.accessor(
+        (row) => (
+          <VStack>
+            <div>{row.encrypted ? 'Yes' : 'No'}</div>
+            {row.encrypted && (
+              <VStack className={styles.cellStack}>
+                <Box className={styles.cellValue}>{row.encryptionAlgo}</Box>
+                <Box className={styles.cellValue}>{row.encryptionKey}</Box>
+              </VStack>
+            )}
+          </VStack>
+        ),
+        {
+          cell: (info) => info.getValue(),
+          header: 'Encryption Details',
+          id: 'encryption'
+        }
+      ),
+      columnHelper.accessor(
+        (row) => (
+          <VStack className={styles.cellStack}>
+            <Box className={styles.cellValue}>{`File: ${row.binLogFileName}`}</Box>
+            <Box className={styles.cellValue}>{`Pos: ${row.binLogFilePos}`}</Box>
+            <Box className={styles.cellValue}>{`GTID: ${row.binLogUuid}`}</Box>
+          </VStack>
+        ),
+        {
+          cell: (info) => info.getValue(),
+          header: 'BinLog Info',
+          id: 'binLogInfo'
+        }
+      ),
+      columnHelper.accessor((row) => row.retentionDays, {
+        cell: (info) => info.getValue(),
+        header: 'Retention (Days)',
+        id: 'retention'
+      }),
+      columnHelper.accessor((row) => (row.completed ? 'Yes' : 'No'), {
+        cell: (info) => info.getValue(),
+        header: 'Completed',
+        id: 'completed'
+      })
+    ],
+    []
+  )
+
+  const backupDataStats = useMemo(
+    () => [
+      {
+        key: 'Total Size',
+        value: backupStats?.total_size
+      },
+      {
+        key: 'Total File Count',
+        value: backupStats?.total_file_count
+      },
+      {
+        key: 'Total Blob Count',
+        value: backupStats?.total_blob_count
+      }
+    ],
+    [backupStats]
+  )
+
+  return (
+    <VStack className={styles.snapshotContainer}>
+      <TableType3 dataArray={backupDataStats} className={styles.statsTable} />
+      <DataTable key="backups" data={data} columns={columns} className={styles.table} />
+    </VStack>
+  )
+}
+
+export default BackupTables
