+++ share/dashboard_react/src/Pages/Maintenance/index.jsx
@@ -1,125 +1,19 @@
-import { createColumnHelper } from '@tanstack/react-table'
-import React, { useEffect, useMemo, useState } from 'react'
-import { convertObjectToArray, formatBytes, formatDate, getBackupMethod, getBackupStrategy } from '../../utility/common'
+import React, { useEffect, useState } from 'react'
+import { convertObjectToArray } from '../../utility/common'
 import AccordionComponent from '../../components/AccordionComponent'
-import { DataTable } from '../../components/DataTable'
 import styles from './styles.module.scss'
-import { Box, HStack, useDisclosure, VStack } from '@chakra-ui/react'
-import TableType3 from '../../components/TableType3'
+import { useDisclosure, VStack } from '@chakra-ui/react'
 import { useDispatch, useSelector } from 'react-redux'
 import BackupSettings from '../Settings/BackupSettings'
 import SchedulerSettings from '../Settings/SchedulerSettings'
 import { TaskLogs } from '../Dashboard/components/Logs'
 import DatabaseJobs from './DatabaseJobs'
 import { purgeResticSnapshot, resticQueueCancel, resticQueueMove, resticQueuePause, resticQueueResume } from '../../redux/clusterSlice'
-import RMIconButton from '../../components/RMIconButton'
 import ConfirmModal from '../../components/Modals/ConfirmModal'
-import { HiPause, HiPlay, HiTrash } from 'react-icons/hi'
 import { showWarningToast } from '../../redux/toastSlice'
-
-const QueueMoveForm = React.memo(({ list = [], currentId, onChange = (dir, afterId) => { } }) => {
-  const [direction, setDirection] = useState('first');
-
-  const handleDirectionChange = (e) => {
-    setDirection(e.target.value);
-    if (e.target.value !== 'after') {
-      onChange(e.target.value, null);
-    }
-  };
-
-  const handleAfterChange = (e) => {
-    onChange('after', e.target.value);
-  };
-  return (
-    <VStack>
-      <HStack>
-        <Box>Move {direction}:</Box>
-        <Select onChange={handleDirectionChange} value={direction}>
-          <option value="first">First</option>
-          <option value="before">After</option>
-          <option value="last">Last</option>
-        </Select>
-      </HStack>
-      {direction === 'after' && (
-        <HStack>
-          <Box>Move After:</Box>
-          <Select onChange={handleAfterChange}>
-            {list.filter(item => item.task_id !== currentId).map(item => (
-              <option key={item.task_id} value={item.task_id}>Task ID #{item.task_id}</option>
-            ))}
-          </Select>
-        </HStack>
-      )}
-    </VStack>
-  )
-})
-
-const formConfig = {
-  queueMove: {
-    component: QueueMoveForm,
-    getProps: (ctx) => ({
-      list: ctx.queueData,
-      currentId: ctx.payload.data.taskId,
-      onChange: ctx.handleMove
-    })
-  }
-};
-
-const DynamicForm = (ctx) => {
-  const { action } = ctx.payload;
-
-  const entry = formConfig[action];
-  if (!entry) return null;
-
-  const Component = entry.component;
-  const props = entry.getProps(ctx);
-
-  return <Component {...props} />;
-};
-
-const resticTaskType = (rtt) => {
-  switch (rtt) {
-    case 0:
-      return "init"
-    case 1:
-      return "fetch"
-    case 2:
-      return "backup"
-    case 3:
-      return "purge"
-    case 4:
-      return "unlock"
-    case 5:
-      return "changepass"
-    default:
-      return "Unknown"
-  }
-}
-
-const resticTaskDetail = (row) => {
-  switch (row.task_type) {
-    case 2:
-      return (<VStack>
-        <HStack>
-          <Box>Path:</Box>
-          <Box>{row.dir_path}</Box>
-        </HStack>
-        <HStack>
-          <Box>Tags:</Box>
-          <Box>{row.tags?.join(', ')}</Box>
-        </HStack>
-      </VStack>)
-    case 3:
-      return (<VStack>
-        <HStack>
-          <Box>Options:</Box>
-          <Box>{JSON.stringify(row.opt)}</Box>
-        </HStack>
-      </VStack>)
-    default:
-      return (<div>-</div>)
-  }
-}
+import BackupTables from './components/BackupTables'
+import SnapshotTables from './components/SnapshotTables'
+import ConfirmActionForm from './components/ConfirmActionForm'
 
 
 function Maintenance({ selectedCluster, user }) {
@@ -130,7 +24,6 @@ function Maintenance({ selectedCluster, user }) {
   const { isOpen: isConfirmModalOpen, title, payload } = confirmState
 
   const dispatch = useDispatch()
-  const columnHelper = createColumnHelper()
   const { isOpen: isBackupSettingsOpen, onToggle: onBackupSettingsToggle } = useDisclosure({
     defaultIsOpen: JSON.parse(localStorage.getItem('isBackupSettingsOpen')) || false
   })
@@ -252,215 +145,6 @@ function Maintenance({ selectedCluster, user }) {
     }
   }, [selectedCluster?.name, resticQueue])
 
-  const backupDataStats = [
-    {
-      key: 'Total Size',
-      value: backupStats?.total_size
-    },
-    {
-      key: 'Total File Count',
-      value: backupStats?.total_file_count
-    },
-    {
-      key: 'Total Blob Count',
-      value: backupStats?.total_blob_count
-    }
-  ]
-
-  const columns = useMemo(
-    () => [
-      columnHelper.accessor((row) => row.id, {
-        cell: (info) => info.getValue(),
-        header: 'ID',
-        id: 'id'
-      }),
-      columnHelper.accessor(
-        (row) => (
-          <>
-            {formatDate(row.startTime)} <br />
-            {formatDate(row.endTime)}
-          </>
-        ),
-        {
-          cell: (info) => info.getValue(),
-          header: 'Start - End Time',
-          id: 'startendTime',
-          minWidth: 160
-        }
-      ),
-      columnHelper.accessor(
-        (row) => (
-          <VStack className={styles.cellStack}>
-            <Box className={styles.cellValue}>{getBackupMethod(row.backupMethod)}</Box>
-            <Box className={styles.cellValue}>{row.backupTool}</Box>
-          </VStack>
-        ),
-        {
-          cell: (info) => info.getValue(),
-          header: 'Backup Method / Tool',
-          id: 'backupMethod'
-        }
-      ),
-      columnHelper.accessor((row) => getBackupStrategy(row.backupStrategy), {
-        cell: (info) => info.getValue(),
-        header: 'Strategy',
-        id: 'strategy'
-      }),
-      columnHelper.accessor(
-        (row) => (
-          <VStack className={styles.cellStack}>
-            <Box className={styles.cellValue}>{row.source}</Box>
-            <Box className={styles.cellValue}>{row.dest}</Box>
-          </VStack>
-        ),
-        {
-          cell: (info) => info.getValue(),
-          header: 'Source - Dest',
-          id: 'srcDest'
-        }
-      ),
-      columnHelper.accessor((row) => formatBytes(row.size), {
-        cell: (info) => info.getValue(),
-        header: 'Backup Size',
-        id: 'backupSize',
-        minWidth: 100
-      }),
-      columnHelper.accessor((row) => (row.compressed ? 'Yes' : 'No'), {
-        cell: (info) => info.getValue(),
-        header: 'Compressed',
-        id: 'compression'
-      }),
-      columnHelper.accessor(
-        (row) => (
-          <VStack>
-            <div>{row.encrypted ? 'Yes' : 'No'}</div>
-            {row.encrypted && (
-              <VStack className={styles.cellStack}>
-                <Box className={styles.cellValue}>{row.encryptionAlgo}</Box>
-                <Box className={styles.cellValue}>{row.encryptionKey}</Box>
-              </VStack>
-            )}
-          </VStack>
-        ),
-        {
-          cell: (info) => info.getValue(),
-          header: 'Encryption Details',
-          id: 'encryption'
-        }
-      ),
-      columnHelper.accessor(
-        (row) => (
-          <VStack className={styles.cellStack}>
-            <Box className={styles.cellValue}>{`File: ${row.binLogFileName}`}</Box>
-            <Box className={styles.cellValue}>{`Pos: ${row.binLogFilePos}`}</Box>
-            <Box className={styles.cellValue}>{`GTID: ${row.binLogUuid}`}</Box>
-          </VStack>
-        ),
-        {
-          cell: (info) => info.getValue(),
-          header: 'BinLog Info',
-          id: 'binLogInfo'
-        }
-      ),
-      columnHelper.accessor((row) => row.retentionDays, {
-        cell: (info) => info.getValue(),
-        header: 'Retention (Days)',
-        id: 'retention'
-      }),
-      columnHelper.accessor((row) => (row.completed ? 'Yes' : 'No'), {
-        cell: (info) => info.getValue(),
-        header: 'Completed',
-        id: 'completed'
-      })
-    ]
-  )
-
-  const snapshotDataStats = [
-    {
-      key: 'Total Size',
-      value: stats?.total_size
-    },
-    {
-      key: 'Total File Count',
-      value: stats?.total_file_count
-    },
-    {
-      key: 'Total Blob Count',
-      value: stats?.total_blob_count
-    }
-  ]
-
-  const snapshotColumns = useMemo(() => [
-    columnHelper.accessor((row) => row.short_id, {
-      header: 'ID',
-      id: 'id'
-    }),
-    columnHelper.accessor((row) => row.time, {
-      header: 'Time'
-    }),
-    columnHelper.accessor((row) => row.paths?.join(','), {
-      header: 'Path'
-    }),
-    columnHelper.accessor((row) => row.hostname, {
-      header: 'Hostname'
-    }),
-    columnHelper.accessor((row) => row.tags?.join(','), {
-      header: 'Tags'
-    }),
-    // Added Purge action column
-    columnHelper.accessor((row) => (
-      <RMIconButton icon={HiTrash} onClick={() => openConfirmModal('Purge Snapshot', { action: 'snapshotPurge', data: { snapshotId: row.id } })} />
-    ), {
-      cell: (info) => info.getValue(),
-      header: 'Actions',
-      id: 'actions',
-    })
-  ])
-
-  const queuelength = queueData?.length || 0;
-
-  const queueDataHeader = useMemo(() =>[
-    {
-      key: 'Total Pending Tasks',
-      value: queuelength
-    },
-    {
-      key: 'Queue Status',
-      value: selectedCluster?.isResticQueuePaused ? 'Paused' : 'Running'
-    },
-    {
-      key: 'Action',
-      value: (selectedCluster?.isResticQueuePaused ?
-        <RMIconButton icon={HiPlay} onClick={() => openConfirmModal('Resume Restic Queue', { action: 'queueResume' })} /> :
-        <RMIconButton icon={HiPause} onClick={() => openConfirmModal('Pause Restic Queue', { action: 'queuePause' })} />
-      )
-    }
-  ], [selectedCluster?.isResticQueuePaused, queuelength])
-
-  const queueColumns = useMemo(() => [
-    columnHelper.accessor((row) => row.task_id, {
-      header: 'ID',
-      id: 'task_id'
-    }),
-    columnHelper.accessor((row) => resticTaskType(row.task_type), {
-      header: 'Task Type'
-    }),
-    columnHelper.accessor((row) => resticTaskDetail(row), {
-      header: 'Details',
-      cell: (info) => info.getValue(),
-      id: 'details',
-      minWidth: 200
-    }),
-    // Added Purge action column
-    columnHelper.accessor((row) => (
-      <RMIconButton icon={HiTrash} onClick={() => openConfirmModal('Cancel Queued Task', { action: 'queueCancel', data: { taskId: row.task_id } })} />
-    ), {
-      cell: (info) => info.getValue(),
-      header: 'Actions',
-      id: 'actions',
-    })
-  ])
-
   return (
     <VStack className={styles.backupContainer}>
       <AccordionComponent
@@ -489,10 +173,7 @@ function Maintenance({ selectedCluster, user }) {
         headerClassName={styles.accordionHeader}
         panelClassName={styles.accordionPanel}
         body={
-          <VStack className={styles.snapshotContainer}>
-            <TableType3 dataArray={backupDataStats} className={styles.statsTable} />
-            <DataTable key="backups" data={data} columns={columns} className={styles.table} />
-          </VStack>
+          <BackupTables data={data} backupStats={backupStats} />
         }
       />
       <AccordionComponent
@@ -503,12 +184,13 @@ function Maintenance({ selectedCluster, user }) {
         headerClassName={styles.accordionHeader}
         panelClassName={styles.accordionPanel}
         body={
-          <VStack className={styles.snapshotContainer}>
-            <TableType3 dataArray={snapshotDataStats} className={styles.statsTable} />
-            <DataTable key="snapshot" data={snapshotData} columns={snapshotColumns} className={styles.table} />
-            <TableType3 dataArray={queueDataHeader} className={styles.statsTable} />
-            <DataTable key="queue" data={queueData} columns={queueColumns} className={styles.table} />
-          </VStack>
+          <SnapshotTables
+            snapshotData={snapshotData}
+            snapshotStats={stats}
+            queueData={queueData}
+            isQueuePaused={selectedCluster?.isResticQueuePaused}
+            onConfirmAction={openConfirmModal}
+          />
         }
       />
       <AccordionComponent
@@ -529,12 +211,21 @@ function Maintenance({ selectedCluster, user }) {
         heading={'Job Logs'}
         body={<TaskLogs />}
       />
-      {isConfirmModalOpen && <ConfirmModal title={title} isOpen={isConfirmModalOpen} body={<DynamicForm
-        payload={payload}
-        queueData={queueData}
-        selectedCluster={selectedCluster}
-        handleMove={handleMove}
-      />} onConfirmClick={handleConfirm} closeModal={closeConfirmModal} />}
+      {isConfirmModalOpen && (
+        <ConfirmModal
+          title={title}
+          isOpen={isConfirmModalOpen}
+          body={
+            <ConfirmActionForm
+              payload={payload}
+              queueData={queueData}
+              onMove={handleMove}
+            />
+          }
+          onConfirmClick={handleConfirm}
+          closeModal={closeConfirmModal}
+        />
+      )}
     </VStack>
   )
 }
